Move EmptyNSMMonitorHandler to the new file -- empty_nsm_monitor_handler.go
Add empty realization of method Update from Handler interface to EmptyNSMMonitorHandler

Add NewNsmEmptyMonitorHandler function to empty_nsm_monitor_handler.go
